refactor: extract functionality of Progress bar
